laba3: изменения в Point

+++ laba_3/include/point.hpp
@@ -3,21 +3,21 @@
 class Point {
 public:
     Point(): __x(0), __y(0) {}
-    Point(const float x, const float y): __x(x), __y(y) {}
+    Point(const double x, const double y): __x(x), __y(y) {}
     ~Point() {}
 
-    float x() const {return __x;}
-    float y() const {return __y;}
-    float len() const;
+    double x() const {return __x;}
+    double y() const {return __y;}
+    double len() const;
 
     Point operator+(const Point& rhs) const;
     Point operator-(const Point& rhs) const;
-    float operator*(const Point& rhs) const;
+    double operator*(const Point& rhs) const;
 
     bool operator==(const Point& rhs) const;
     bool operator!=(const Point& rhs) const {return !(*this == rhs);}
     
 private:
-    float __x;
-    float __y;
+    double __x;
+    double __y;
 };
