+++ laba_3/src/point.cpp
@@ -1,6 +1,6 @@
 #include "../include/point.hpp"
 
-float Point::len() const {
+double Point::len() const {
     return __x * __x + __y * __y;
 }
 
@@ -12,7 +12,7 @@ Point Point::operator-(const Point& rhs) const {
     return Point(__x - rhs.__x, __y - rhs.__y);
 }
 
-float Point::operator*(const Point& rhs) const {
+double Point::operator*(const Point& rhs) const {
     return __x * rhs.__x + __y * rhs.__y;
 }
